Refactor Observable code and update package version
This commit includes a significant refactor of the Observable project to improve and optimize the overall functionality. The Observable code was rewritten to make it cleaner and more efficient. Also, the package version of 'evg_observable' was updated in package.json file. Additional Typescript files such as Types.ts, Observable.ts, etc., have been added for better code management. Existing bugs were addressed and resolved in the process.