Update evg_observable package version in package.json
The Observable project has undergone a significant refactor to enhance functionality and optimisation. The refactoring includes a rewrite of the Observable code for cleaner and more efficient usage. The 'evg_observable' package version has been updated in the package.json file. Other improvements include the addition of Typescript files for better code management and bug resolutions.