Update conditions in detectChanges method
The code in the detectChanges method and detectForCycle function inside the RootHtmlElement.ts file has been modified. The condition check for the ahe_ForOfList length is now also included within the not operator in the detectChanges method. The isCustomAppElement attribute check is moved inside the detectForCycle function for better contextual relevance. These changes improve the logic flow and efficiency within the file.