Add 'qsi-bind' root element and bindings
A new root element 'qsi-bind' is added to the root element enumeration. It is registered in the RegisterRootElements method in 'registerRootElements.ts' and used in 'main.html'. A new file 'qsiBind.ts' is created to accommodate the 'qsi-bind' element. Additional functions and variables are added in 'utils.ts' and 'RootHtmlElement.ts' to handle 'qsi-bind' bindings.