Reorder code and add sorting functionality in Observable classes
This update reorganizes some segments of the code in Observable classes for greater readability and maintainability. The 'get' and 'set' methods for 'isDestroyed', 'isEnable', and 'order' have been relocated. The errorHandler initialization and _order variable declaration were also repositioned. In addition, ascending and descending sorting functionality has been added to the "OrderedObservable" class using the newly added 'sortAscending' and 'sortDescending' functions in the FunctionLibs.