Updated Observable logic and added sorting functionality
The logic in the various Observable classes has been simplified and streamlined, making it more readable and easier to manage. These updates mainly affect 'get' and 'set' methods for 'isDestroyed', 'isEnable', and 'order'. The most significant addition is a new sorting feature in the "OrderedObservable" class, which can sort Observables in ascending or descending order.