Update HTML tags and change variable casting
Updated 'main.html' to modify 'app-child' tags to use a 'qsi-channel' property replacing 'qsi-inject_to'. The 'utils.ts' file was revised to fix assigning a component to the 'channelName' property of the rootElement. Now 'element' is forcibly cast to 'any' before it is cast to 'IChannel'. These changes aim to improve the flexibility and clarity of channel handling.