Introduced pipe and chain functionality in Observables
The central parts of the Observables module are restructured to introduce the concept of pipe and chain functionality. Now, the 'SubscribeObject' class extends 'Pipe' and sets up the foundation for chain methods. Besides, some types and functions are added or modified to accommodate this new feature. A few other cosmetic changes were made as well, such as changing the order functions and better defining the getListener method.