Refactor evg_observable module, add filter functionality to Observable
This commit includes a major refactoring of the evg_observable module. It adds a new "Filter" functionality to the Observable class, allowing more precise control over observables' data processing. It also modifies data subscription mechanisms to support subscribing as an Array of observers, resulting in more flexible and robust usage.