Update evg_observable version and adjust test_obsevrable.js
The commit includes an update to the version of evg_observable package in package.json. Additionally, there are extensive changes made to the builds/test_obsevrable/test_obsevrable.js, including adjusting the test script. These improvements aim to optimize the performance of the script and to make the code more readable.